♻️ control-plane: collapse vestigial provisioner contract to a tier gate (DOMAIN.T3)
Operator owns provisioning, so delete the dead ClusterTenantProvisioner interface +
its shared/external-webhook implementations (provision/getStatus/deprovision were
zero-runtime-call) and the unused provisionerFor registry method. Registry is now a
pure {id,tiers} tier-availability gate (isTierAvailable + capabilities). dedicatedCluster
gating + HTTPS-only webhook-config validation preserved. provisioner.test→registry.test.
