--- apps/control-plane/src/core/cluster-tenants/provisioner.ts
@@ -1,3 +1,0 @@
-export { SHARED_PROVISIONER_ID, SharedClusterProvisioner } from "./shared-cluster.provisioner.js";
-export { ExternalWebhookProvisioner } from "./external-webhook.provisioner.js";
-export { _ReadExternalWebhookConfig } from "./external-webhook.config.js";
