+++ apps/control-plane/src/core/cluster-tenants/registry.infra.ts
@@ -1,61 +1,50 @@
 import { ClusterTenantIsolationTier } from "@opencrane/contracts";
 import type { ClusterTenantProvisionerCapability, ClusterTenantProvisionerRegistry } from "@opencrane/contracts";
 
-import type { ClusterTenantProvisioner } from "./provisioner.types.js";
-
 /**
- * A registered provisioner paired with the tiers it advertises.
+ * A registered backend: just its stable id and the tiers it advertises.
+ *
+ * The control plane no longer holds provisioner *lifecycle* logic — the operator's
+ * ClusterTenant reconciler owns provision/deprovision (see DOMAIN.T1/T2). All this
+ * registry does at runtime is answer "can a backend serve this tier?", so an entry
+ * needs nothing more than `{ id, tiers }`.
  */
 export interface RegisteredProvisioner
 {
   /** Stable provisioner identifier surfaced in capabilities. */
   id: string;
-  /** Isolation tiers this provisioner can satisfy. */
+  /** Isolation tiers this provisioner advertises. */
   tiers: ClusterTenantIsolationTier[];
-  /** The provisioner instance that materialises the boundary. */
-  provisioner: ClusterTenantProvisioner;
 }
 
 /**
- * Concrete registry holding the built-in shared provisioner plus any external
- * webhook backend. Tier→provisioner routing is by advertised capability, so the
- * management API can ask whether a tier can be served and resolve the backend
- * that owns it without knowing about specific vendors.
+ * Tier-availability gate for the management API. Tells the create/update path
+ * whether some registered backend advertises a requested isolation tier, and
+ * advertises the full capability set — nothing more. The provisioning itself is
+ * the operator's job.
  */
 export class DefaultClusterTenantProvisionerRegistry implements ClusterTenantProvisionerRegistry
 {
-  /** Every registered provisioner with its advertised tiers, in routing order. */
+  /** Every registered backend with its advertised tiers. */
   private readonly registered: RegisteredProvisioner[];
 
   /**
-   * @param entries - Pre-built provisioner registrations (built-in first, external last).
+   * @param entries - Pre-built registrations (built-in first, external last).
    */
   constructor(entries: RegisteredProvisioner[])
   {
     this.registered = entries;
   }
 
-  /** Whether some registered provisioner advertises the given isolation tier. */
+  /** Whether some registered backend advertises the given isolation tier. */
   isTierAvailable(tier: ClusterTenantIsolationTier): boolean
   {
     return this.registered.some(entry => entry.tiers.includes(tier));
   }
 
-  /** Capabilities of every registered provisioner. */
+  /** Capabilities of every registered backend. */
   capabilities(): ClusterTenantProvisionerCapability[]
   {
     return this.registered.map(entry => ({ id: entry.id, supportedTiers: entry.tiers }));
-  }
-
-  /**
-   * Resolve the provisioner that owns the given tier.
-   *
-   * @param tier - Isolation tier to route.
-   * @returns The owning provisioner, or null when no backend serves the tier.
-   */
-  provisionerFor(tier: ClusterTenantIsolationTier): ClusterTenantProvisioner | null
-  {
-    const entry = this.registered.find(item => item.tiers.includes(tier));
-    return entry ? entry.provisioner : null;
   }
 }
