+++ apps/control-plane/src/core/cluster-tenants/registry.ts
@@ -1,39 +1,42 @@
 import { ClusterTenantIsolationTier } from "@opencrane/contracts";
 
-import { SHARED_PROVISIONER_ID, SharedClusterProvisioner } from "./shared-cluster.provisioner.js";
-import { ExternalWebhookProvisioner } from "./external-webhook.provisioner.js";
 import { _ReadExternalWebhookConfig } from "./external-webhook.config.js";
 import { DefaultClusterTenantProvisionerRegistry } from "./registry.infra.js";
 import type { RegisteredProvisioner } from "./registry.infra.js";
 
+/** Stable identifier the built-in shared backend advertises in the registry. */
+export const SHARED_PROVISIONER_ID = "shared";
+
 /**
  * Build the default registry from the environment.
  *
- * @returns A registry advertising the built-in shared tiers, plus the external
- *   webhook backend (advertising `dedicatedCluster`) only when configured.
+ * The registry is now a pure tier-availability gate (the operator owns provisioning,
+ * see DOMAIN.T1/T2), so it carries only `{ id, tiers }` advertisements:
+ *  - the built-in shared backend always advertises the two in-cluster tiers;
+ *  - the external webhook backend advertises `dedicatedCluster` only when its env is
+ *    configured (and validated as HTTPS), so the tier stays unavailable — rejected by
+ *    the API — on installs that have not opted into an out-of-process backend.
+ *
+ * @returns A registry advertising the built-in shared tiers, plus `dedicatedCluster`
+ *   when an external webhook backend is configured.
  */
 export function _BuildClusterTenantProvisionerRegistry(): DefaultClusterTenantProvisionerRegistry
 {
-  // 1. Always register the built-in shared provisioner — it serves the two
-  //    in-cluster tiers and is the platform's default isolation backend.
   const entries: RegisteredProvisioner[] = [
     {
       id: SHARED_PROVISIONER_ID,
       tiers: [ClusterTenantIsolationTier.Shared, ClusterTenantIsolationTier.DedicatedNodes],
-      provisioner: new SharedClusterProvisioner(),
     },
   ];
 
-  // 2. Register the external webhook backend only when its env is configured, so
-  //    `dedicatedCluster` stays unavailable (rejected by the API) on installs
-  //    that have not opted into an out-of-process provisioner.
+  // `_ReadExternalWebhookConfig` still gates the tier (and fails loud on a non-HTTPS
+  // endpoint), even though the control plane no longer POSTs to the backend itself.
   const externalConfig = _ReadExternalWebhookConfig();
   if (externalConfig)
   {
     entries.push({
       id: externalConfig.id,
       tiers: [ClusterTenantIsolationTier.DedicatedCluster],
-      provisioner: new ExternalWebhookProvisioner(externalConfig),
     });
   }
